expose wandb.controller
